Add budget enforcement with a kernel-side wall clock (T1.2.3)
The SDK enforces per-session cost and turn limits, but a task may run
several sessions when output fails validation. Only the kernel sees the
total, so the task-level bound lives here: one ledger per task, shared
across retries, and each retry is offered the budget that is left rather
than the whole of it. Three sessions that each stay under the limit can
still blow through it together.

The wall-clock timer does not trust the session to honour its abort
signal. A wedged subprocess or a socket that never closes would
otherwise stall the run, so the kernel stops waiting on its own schedule
and reports wall_clock regardless of what the session does next. The
hung-session test uses a provider that ignores abort entirely, which is
the case the timer exists for.

A session that stops on its own budget and a ledger that detects a
breach must not both log the same stop, so the termination check settles
first and returns before the ledger check runs.

wallClockSeconds is no longer constrained to whole seconds. Sub-second
budgets are legitimate, and they let the timer be tested in
milliseconds rather than by making the suite wait.

Traces: DESIGN §4.1, AGT-4.

+++ src/role/definition.ts
@@ -34,7 +34,7 @@ export const budgetSchema = z.object({
   tokens: z.number().int().positive(),
   costUsd: z.number().positive(),
   steps: z.number().int().positive(),
-  wallClockSeconds: z.number().int().positive(),
+  wallClockSeconds: z.number().positive(),
 });
 
 export const outputSchema = z.object({
